fix: update unittest
