+++ .gitignore
@@ -25,3 +25,4 @@ output
 /.idea/
 dependency-reduced-pom.xml
 grpc-dummy-server.log
+jmeter.log
