+++ src/test/java/vn/zalopay/benchmark/core/ClientCallerTest.java
@@ -8,47 +8,16 @@ import io.netty.handler.ssl.SslContextBuilder;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import vn.zalopay.benchmark.core.protobuf.ProtocInvoker;
 import vn.zalopay.benchmark.core.specification.GrpcResponse;
 
 import javax.net.ssl.SSLException;
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 import static org.mockito.Mockito.any;
 
-public class ClientCallerTest {
-    private static final String GRPC_DUMMY_SERVER_JAR = "gprc-server-1.0-SNAPSHOT.jar";
-    private static final Path GRPC_DUMMY_SERVER_FOLDER = Paths.get(System.getProperty("user.dir"), "/dist/benchmark/grpc-server/dist");
-    private static final Path PROTO_WITH_EXTERNAL_IMPORT_FOLDER =
-            Paths.get(System.getProperty("user.dir"), "dist/benchmark/grpc-server/src/main/resources/protos-v2");
-    private static final Path LIB_FOLDER =
-            Paths.get(System.getProperty("user.dir"), "dist/benchmark/grpc-server/src/main/resources/libs");
+public class ClientCallerTest extends BaseTest {
 
-    private static String HOST_PORT = "localhost:8005";
-    private static String HOST_PORT_TLS = "localhost:8006";
-    private static String REQUEST_JSON = "{\"shelf\":{\"id\":1599156420811,\"theme\":\"Hello server!!\"}}";
-    private static String FULL_METHOD = "bookstore.Bookstore/CreateShelf";
-    private static String METADATA = "";
-    private ClientCaller clientCaller;
-    static Process bookStoreServer;
-    static Process helloWorldServer;
-    static Process bookStoreTlsServer;
-    static Process helloWorldTlsServer;
-    static File dummyLog;
-
-    @BeforeClass
-    public void setupDependencies() throws IOException {
-        System.setProperty("javax.net.ssl.trustStore", Paths.get(System.getProperty("user.dir"), "dist", "cert", "cacert").toString());
-        System.setProperty("javax.net.ssl.trustStorePassword", "changeit");
-        startDummyGrpcServer();
-    }
 
     @Test
     public void testCanSendGrpcUnaryRequest() {
@@ -223,7 +192,7 @@ public class ClientCallerTest {
     public void testThrowExceptionWithTimeoutRequest() {
         clientCaller = new ClientCaller(HOST_PORT, PROTO_WITH_EXTERNAL_IMPORT_FOLDER.toString(), LIB_FOLDER.toString(), FULL_METHOD, false, false, METADATA);
         clientCaller.buildRequest(REQUEST_JSON);
-        GrpcResponse resp = clientCaller.call("10");
+        GrpcResponse resp = clientCaller.call("1");
         clientCaller.shutdownNettyChannel();
         Assert.assertNotNull(resp);
         Assert.assertTrue(resp.getGrpcMessageString().contains("\"theme\": \"Hello server"));
@@ -233,7 +202,7 @@ public class ClientCallerTest {
     public void testThrowExceptionWithTimeoutRequestServerStream() {
         clientCaller = new ClientCaller(HOST_PORT, PROTO_WITH_EXTERNAL_IMPORT_FOLDER.toString(), LIB_FOLDER.toString(), FULL_METHOD, false, false, METADATA);
         clientCaller.buildRequest(REQUEST_JSON);
-        GrpcResponse resp = clientCaller.callServerStreaming("10");
+        GrpcResponse resp = clientCaller.callServerStreaming("1");
         clientCaller.shutdownNettyChannel();
         Assert.assertNotNull(resp);
         Assert.assertTrue(resp.getGrpcMessageString().contains("\"theme\": \"Hello server"));
@@ -243,7 +212,7 @@ public class ClientCallerTest {
     public void testThrowExceptionWithTimeoutRequestClientStream() {
         clientCaller = new ClientCaller(HOST_PORT, PROTO_WITH_EXTERNAL_IMPORT_FOLDER.toString(), LIB_FOLDER.toString(), FULL_METHOD, false, false, METADATA);
         clientCaller.buildRequest(REQUEST_JSON);
-        GrpcResponse resp = clientCaller.callClientStreaming("10");
+        GrpcResponse resp = clientCaller.callClientStreaming("1");
         clientCaller.shutdownNettyChannel();
         Assert.assertNotNull(resp);
         Assert.assertTrue(resp.getGrpcMessageString().contains("\"theme\": \"Hello server"));
@@ -325,49 +294,5 @@ public class ClientCallerTest {
     public void testThrowExceptionWithInvalidServiceName() {
         clientCaller = new ClientCaller(HOST_PORT, PROTO_WITH_EXTERNAL_IMPORT_FOLDER.toString(), LIB_FOLDER.toString(), "bookstore.Bookstores/CreateShelf", false, false, METADATA);
         clientCaller.buildRequest(REQUEST_JSON);
-    }
-
-
-    @AfterMethod
-    public void cleanMockito() {
-        Mockito.clearAllCaches();
-        if (clientCaller != null)
-            clientCaller.shutdownNettyChannel();
-        clientCaller = null;
-    }
-
-    @AfterClass
-    public void shutdownDummyServer() throws InterruptedException {
-        bookStoreServer.destroy();
-        if (bookStoreServer.isAlive())
-            bookStoreServer.destroyForcibly();
-        helloWorldServer.destroy();
-        if (helloWorldServer.isAlive())
-            helloWorldServer.destroyForcibly();
-        bookStoreTlsServer.destroy();
-        if (bookStoreTlsServer.isAlive())
-            bookStoreTlsServer.destroyForcibly();
-        helloWorldTlsServer.destroy();
-        if (helloWorldTlsServer.isAlive())
-            helloWorldTlsServer.destroyForcibly();
-    }
-
-    private void startDummyGrpcServer() throws IOException {
-        File javaHome = new File(System.getProperty("java.home"), "bin");
-        String javaPath = javaHome + File.separator + "java";
-        String startClassPathCommand = String.format("%s", Paths.get(GRPC_DUMMY_SERVER_FOLDER.toString(), GRPC_DUMMY_SERVER_JAR.toString()).toString());
-        ProcessBuilder startBookStoreGRPCServerProcessBuilder = new ProcessBuilder(javaPath, "-cp", startClassPathCommand, "server.BookStoreServer");
-        ProcessBuilder startHelloWorldGRPCServerProcessBuilder = new ProcessBuilder(javaPath, "-cp", startClassPathCommand, "server.HelloWorldServer");
-        ProcessBuilder startBookStoreTlsGRPCServerProcessBuilder = new ProcessBuilder(javaPath, "-cp", startClassPathCommand, "server.BookStoreServerTls");
-        ProcessBuilder startHelloWorldTlsGRPCServerProcessBuilder = new ProcessBuilder(javaPath, "-cp", startClassPathCommand, "server.HelloWorldServerTls");
-        dummyLog = new File("grpc-dummy-server.log");
-        startBookStoreGRPCServerProcessBuilder.redirectError(ProcessBuilder.Redirect.appendTo(dummyLog));
-        startHelloWorldGRPCServerProcessBuilder.redirectError(ProcessBuilder.Redirect.appendTo(dummyLog));
-        startBookStoreTlsGRPCServerProcessBuilder.redirectError(ProcessBuilder.Redirect.appendTo(dummyLog));
-        startHelloWorldTlsGRPCServerProcessBuilder.redirectError(ProcessBuilder.Redirect.appendTo(dummyLog));
-        bookStoreServer = startBookStoreGRPCServerProcessBuilder.start();
-        helloWorldServer = startHelloWorldGRPCServerProcessBuilder.start();
-        bookStoreTlsServer = startBookStoreTlsGRPCServerProcessBuilder.start();
-        helloWorldTlsServer = startHelloWorldTlsGRPCServerProcessBuilder.start();
     }
 }
